jdbcTemplate excute return 타입은 꼭 Long이 아닐수도 있다

+++ src/main/java/persistence/core/DefaultEntityPersister.java
@@ -18,7 +18,7 @@ public class DefaultEntityPersister implements EntityPersister {
     @Override
     public Long insert(Object entity) {
         String sql = dmlQueryBuilder.insertSql(entity);
-        return jdbcTemplate.execute(sql);
+        return (Long) jdbcTemplate.execute(sql);
     }
 
     @Override
